fix: perps socket subscription
