+++ development/webpack/utils.js
@@ -3,9 +3,7 @@ const { developmentConsts } = require('../babelTools');
 const { EXT_CHANNEL, TARGET_BROWSER } = require('./constant');
 
 exports.createResolveExtensions = function ({ platform, configName }) {
-  console.log('createResolveExtensions>>>>>>', platform, configName);
-
-  return uniq([
+  const result = uniq([
     // .chrome-ext.ts, .firefox-ext.ts
     ...(EXT_CHANNEL && TARGET_BROWSER
       ? ['.ts', '.tsx', '.js', '.jsx'].map(
@@ -41,6 +39,8 @@ exports.createResolveExtensions = function ({ platform, configName }) {
     '.wasm',
     '.d.ts',
   ]);
+  console.log('createResolveExtensions>>>>>>', platform, configName, result);
+  return result;
 };
 
 exports.getOutputFolder = function () {
